requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
torch-scatter>=2.1.0

# 配置管理
hydra-core>=1.3.0
omegaconf>=2.3.0

# 数据处理
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0

# 可视化
matplotlib>=3.6.0
seaborn>=0.12.0

# 工具库
tqdm>=4.65.0
pyyaml>=6.0

# 可选：实验跟踪
# tensorboard>=2.12.0
# wandb>=0.15.0
